refactor: 프로젝트 등록 모달 오픈 로직 분리 (#174)
