+++ src/components/domain/project/projectButton/ProjectRegisterButton.tsx
@@ -4,11 +4,7 @@ import { cn } from '@/lib/utils';
 import { ClipboardEdit } from 'lucide-react';
 import { Button } from '@/components/ui/button';
 
-import { useModalStore } from '@/stores/modalStore';
-import { useAuthStore } from '@/stores/authStore';
-
-import LoginModal from '../../auth/login/LoginModal';
-import ProjectRegisterModal from '@/components/domain/project/projectRegisterModal/ProjectRegisterModal';
+import useProjectRegisterModal from '../hooks/useProjectRegisterModal';
 
 interface ProjectRegisterButtonProps {
   text?: string;
@@ -19,17 +15,7 @@ export default function ProjectRegisterButton({
   text = '내 프로젝트 등록하기',
   className,
 }: ProjectRegisterButtonProps) {
-  const openModal = useModalStore((state) => state.openModal);
-
-  const isLoggedIn = useAuthStore((state) => state.isLoggedIn);
-
-  const handleOpenProjectRegisterModal = () => {
-    if (!isLoggedIn) {
-      openModal(<LoginModal />);
-      return;
-    }
-    openModal(<ProjectRegisterModal />);
-  };
+  const { handleOpenProjectRegisterModal } = useProjectRegisterModal();
 
   return (
     <Button
